astra/backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
python-dotenv==1.2.2
httpx==0.28.1
openai==2.29.0
sqlalchemy==2.0.48
pydantic==2.12.5
pydantic-settings==2.7.0
aiosqlite==0.22.1
